==> 7_cluster_figs.Rmd <==
---
title: "7_cluster_figs"
author: "phil"
date: "7/17/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache = TRUE)
```

```{r}
map_width <- 4
map_height <- 3
```


```{r}
library(magrittr)
library(tidyverse)
library(scales)
library(ggthemes)
library(ggrepel)
library(igraph)
library(compx)
library(RColorBrewer)
library(gplots)
library(viridis)
library(ggmap)
library(sf)


source('R/utils.R')
```

```{r}
all_data <- read_csv('data/demographics.csv', 
                 col_types = cols(
                     GEOID = col_character()
                     )
                 ) %>% 
    rename(tract = GEOID) %>% 
    group_by(tract) %>% 
    filter(sum(n) > 0) %>% 
    ungroup() 
```

```{r}
df <- data_frame(city = character(), 
                 k = integer(), 
                 j = integer(),
                 hclust = numeric(), 
                 mixed = numeric())
```

```{r}
loss_curves <- data_frame(city = character(),
                          k = integer(),
                          h_val = numeric(), 
                          m_val = numeric())
```

```{r}
MI <- data_frame(city = character(), 
                 I_XY = numeric())
```


# Detroit

```{r}
city <- 'Detroit'
in_data <- read_data(city, all_data)
data    <- in_data$data
tracts  <- in_data$tracts
```

```{r}
g <- construct_information_graph(tracts = tracts, 
								 data = data, 
								 divergence = js_dist)

sigma <- 30
plot_spectrum(g, sigma)

ggsave(paste0('figs/spectrum_', city, '.pdf'))

```

```{r}
k     <- 34
g     <- spectral_cluster(g, sigma = sigma, k = k, nreps = 1000)
```

```{r}
input <- group_data(tracts, data, g) 
blocked_tracts <- input$blocked_tracts
blocked_data <- input$blocked_data
```




## Hierarchical Clustering

```{r}
b <- h_clust(tracts, data)
a <- h_clust(blocked_tracts, blocked_data)
```

```{r}
compare_hclusts(a, b)
```

```{r}

j <- 7

m <- prominence_map(tracts, data) 

ggsave(paste0('figs/prominence_', city, '.pdf'), width = map_width, height = map_height)


s <- s_clust_map(m, tracts, data, blocked_tracts, size_coef = .75)

s

ggsave(paste0('figs/sclust_', city, '.pdf'), width = map_width, height = map_height)


h <- h_clust_map(m, blocked_tracts, blocked_data, a, j, size_coef = .75)
h

ggsave(paste0('figs/mixed_', city,'.pdf'), width = map_width, height = map_height)

```

```{r}

f <- h_clust_map(m, tracts, data, b, j, size_coef = .75)
f
ggsave(paste0('figs/hclust_', city,'.pdf'), width = map_width, height = map_height)
```


```{r}
# total MI
data %>% 
    spread(key = group, value = n) %>% 
    select(-tract) %>% 
    mutual_info()
```

```{r}
df <- df %>% 
    rbind(data_frame(
        city = city, 
        j = j,
        k = k, 
        hclust = cumsum(rev(b$height))[j],
        mixed = cumsum(rev(a$height))[j]
    ))

```

```{r}
loss_curves <- loss_curves %>%
    rbind(data_frame(
        city = city, 
        k = 1:length(a$height),
        h_val = rev(b$height)[1:length(a$height)],
        m_val = rev(a$height)
    ))

MI <- MI %>% 
    rbind(data_frame(
        city = city, 
        I_XY = data %>% 
            spread(key = group, value = n) %>% 
            select(-tract) %>% 
            mutual_info()
    ))
```




### Add place decomposition for first fig

```{r}
detroit <- st_union(tracts)

detroit_places <- tigris::places(c(26), cb = FALSE, year = 2015, refresh = FALSE) %>% 
	st_as_sf()

st_crs(detroit_places) <- st_crs(tracts)

places <- detroit_places %>% 
	mutate(include = st_intersects(detroit_places, detroit),
		   include = map_dbl(include, ~ifelse(length(.) == 0, 0, 1))) %>% 
	filter(include == 1) %>% 
	st_intersection(detroit) %>% 
    mutate(type = st_geometry_type(geometry)) %>%
    filter(type != 'GEOMETRYCOLLECTION')

m + 
	geom_sf(data = places, alpha = 0, size = .5, color = 'black') + 
    theme(legend.text = element_text(size = 8))
	
ggsave('figs/detroit_places.pdf', width = map_width, height = map_height)

```

```{r}
lookup <- st_join(tracts, detroit_places, st_within) %>% 
    select(GEOID.x, NAME.y) %>% 
    select(-geometry)

data %>% 
    left_join(lookup, by = c('tract' = 'GEOID.x')) %>% 
    mutate(NAME = ifelse(is.na(NAME.y), 'NA', NAME.y)) %>% 
    group_by(NAME, group) %>% 
    summarise(n = sum(n)) %>% 
    ungroup() %>% 
    spread(key = group, value = n) %>%
    select(-NAME) %>% 
    mutual_info()

```

```{r}

data %>% 
    spread(key = group, value = n) %>% 
    select(-tract) %>% 
    mutual_info()

```



# Chicago

```{r}
city    <- 'Chicago'
in_data <- read_data(city, all_data)
data    <- in_data$data
tracts  <- in_data$tracts

b <- h_clust(tracts, data)

```

```{r}

g <- construct_information_graph(tracts = tracts, 
								 data = data, 
								 divergence = js_dist)

sigma <- 30
plot_spectrum(g, sigma)
ggsave(paste0('figs/spectrum_', city, '.pdf'))
```

```{r}

set.seed(1250)

k     <- 36
g     <- spectral_cluster(g, sigma = sigma, k = k, nreps = 1000)
```

```{r}
input <- group_data(tracts, data, g) 
blocked_tracts <- input$blocked_tracts
blocked_data <- input$blocked_data

blocked_tracts %>% 
    ggplot() + 
    geom_sf(aes(fill = as.character(cluster))) +
    guides(fill = FALSE)
```

## Hierarchical Clustering

```{r}
a <- h_clust(blocked_tracts, blocked_data)
compare_hclusts(a, b)
```

```{r}

j <- 11

m <- prominence_map(tracts, data) + 
        guides(fill = FALSE) 


s <- s_clust_map(m, tracts, data, blocked_tracts, size_coef = .75)

ggsave(paste0('figs/sclust_', city, '.pdf'), width = map_width, height = map_height)



h <- h_clust_map(m, blocked_tracts, blocked_data, a, j, size_coef = .75)

h

ggsave(paste0('figs/mixed_', city,'.pdf'), width = map_width, height = map_height)
```

```{r}

f <- h_clust_map(m, tracts, data, b, j, size_coef = .75)

f
ggsave(paste0('figs/hclust_', city,'.pdf'), width = map_width, height = map_height)
```

```{r}
df <- df %>% 
    rbind(data_frame(
        city = city, 
        k = k, 
        j = j,
        hclust = cumsum(rev(b$height))[j],
        mixed = cumsum(rev(a$height))[j]
    ))
```


```{r}
loss_curves <- loss_curves %>%
    rbind(data_frame(
        city = city, 
        k = 1:length(a$height),
        h_val = rev(b$height)[1:length(a$height)],
        m_val = rev(a$height)
    ))

MI <- MI %>% 
    rbind(data_frame(
        city = city, 
        I_XY = data %>% 
            spread(key = group, value = n) %>% 
            select(-tract) %>% 
            mutual_info()
    ))
```


# Philadelphia


```{r}
city    <- 'Philadelphia'
in_data <- read_data(city, all_data)
data    <- in_data$data
tracts  <- in_data$tracts

set.seed(1234)

```

## Filter out the small island
```{r}
g <- construct_information_graph(tracts = tracts, 
								 data = data, 
								 divergence = euc)

comps <- igraph::components(g)

largest_cluster <- which.max(table(comps$membership))
exclude <- which(comps$membership != largest_cluster) %>% names()

tracts <- tracts %>% 
    filter(!(GEOID %in% exclude))

data <- data %>%
    filter(tract %in% tracts$GEOID)
```

```{r}
b <- h_clust(tracts, data)
```

```{r}
g <- construct_information_graph(tracts = tracts, 
								 data = data, 
								 divergence = js_dist)

sigma <- 30
plot_spectrum(g, sigma)
ggsave(paste0('figs/spectrum_', city, '.pdf'))
```

```{r}
k     <- 11
g     <- spectral_cluster(g, sigma = sigma, k = k, nreps = 1000)
```

```{r}
input <- group_data(tracts, data, g) 
blocked_tracts <- input$blocked_tracts
blocked_data <- input$blocked_data

blocked_tracts %>% 
    ggplot() + 
    geom_sf(aes(fill = as.character(cluster))) +
    guides(fill = FALSE)
```

## Hierarchical Clustering

```{r}
a <- h_clust(blocked_tracts, blocked_data)
compare_hclusts(a, b)
```

```{r}
j <- 10
m <- prominence_map(tracts, data) + 
        guides(fill = FALSE) 


s <- s_clust_map(m, tracts, data, blocked_tracts, size_coef = .75)

s

ggsave(paste0('figs/sclust_', city, '.pdf'), width = map_width, height = map_height)



h <- h_clust_map(m, blocked_tracts, blocked_data, a, j, size_coef = .75)

h
ggsave(paste0('figs/mixed_', city,'.pdf'), width = map_width, height = map_height)
```

```{r}

f <- h_clust_map(m, tracts, data, b, j, size_coef = .75)

f
ggsave(paste0('figs/hclust_', city,'.pdf'), width = map_width, height = map_height)
```

```{r}
df <- df %>% 
    rbind(data_frame(
        city = city, 
        k = k, 
        j = j,
        hclust = cumsum(rev(b$height))[j],
        mixed = cumsum(rev(a$height))[j]
    ))
```


```{r}
loss_curves <- loss_curves %>%
    rbind(data_frame(
        city = city, 
        k = 1:length(a$height),
        h_val = rev(b$height)[1:length(a$height)],
        m_val = rev(a$height)
    ))

MI <- MI %>% 
    rbind(data_frame(
        city = city, 
        I_XY = data %>% 
            spread(key = group, value = n) %>% 
            select(-tract) %>% 
            mutual_info()
    ))
```

# Loss Curves

```{r}

labels <- data_frame(
    city = c('Detroit', 'Philadelphia', 'Chicago'),
    k    = c(20, 12, 32),
    I_XY   = c(.52, .3, .4)
)

loss_curves %>% 
    group_by(city) %>% 
    mutate(val = cumsum(h_val)) %>% 
    ggplot() + 
    geom_line(aes(x = k, y = val, color = city)) + 
    geom_point(aes(x = k, y = val, color = city)) + 
    theme_bw() + 
    scale_color_brewer(palette = 'Accent') + 
    scale_fill_brewer(palette = 'Accent') + 
    guides(color = FALSE, fill = FALSE) + 
    xlab(expression(italic(k))) + 
    ylab(expression(italic(I(C[k], Y)))) + 
    geom_hline(aes(yintercept = I_XY, color = city), data = MI, linetype = 'dashed', size = 1) + 
    geom_label(aes(x = k, y = I_XY, label = city, fill = city), data = labels) + 
    scale_y_continuous(lim = c(0, NA))

ggsave('figs/loss_curves_h.pdf', width = map_width, height = map_height)
```

```{r}

labels <- data_frame(
    city = c('Detroit', 'Philadelphia', 'Chicago'),
    k    = c(20, 12, 32),
    I_XY   = c(.49, .25, .34)
)

loss_curves %>% 
    group_by(city) %>% 
    mutate(val = cumsum(m_val)) %>% 
    ggplot() + 
    geom_line(aes(x = k, y = val, color = city)) + 
    geom_point(aes(x = k, y = val, color = city)) + 
    theme_bw() + 
    scale_color_brewer(palette = 'Accent') + 
    scale_fill_brewer(palette = 'Accent') + 
    guides(color = FALSE, fill = FALSE) + 
    xlab(expression(italic(k))) + 
    ylab(expression(italic(I(C[k], Y)))) + 
    geom_hline(aes(yintercept = I_XY, color = city), data = MI, linetype = 'dashed', size = 1) + 
    geom_label(aes(x = k, y = I_XY, label = city, fill = city), data = labels) +
    scale_y_continuous(lim = c(0, NA))

ggsave('figs/loss_curves_m.pdf',  width = map_width, height = map_height)
```
